+++ src/output/build.js
@@ -2623,7 +2623,7 @@ class Channel {
                 }
                 if (this.remaining === 0) {
                     task.onCancel();
-                    Messenger.write("Targeted data decryption and <span class='clickable-no-click'>memory core defragmentation</span> complete");
+                    Messenger.write("Targeted data decryption and <span class='clickable-no-click'>memory core decryption</span> complete");
                 }
             }
             else {